Enum 자료형 변경, 펀딩 데이터 타입 변경

+++ src/main/java/com/ia/indieAn/domain/fund/dto/FundEnrollDto.java
@@ -23,7 +23,7 @@ public class FundEnrollDto {
     private Date startDate;
     private Date endDate;
     private Date paymentDate;
-    private int target;
+    private long target;
     private FundStatusEnum fundStatus;
     private List<RewardListDto> reward = new ArrayList<>();
 }
